requirement.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.1.2
django-ckeditor-5==0.2.15
django-import-export==4.2.0
django-js-asset==2.2.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.2
et_xmlfile==2.0.0
idna==3.10
lxml==5.3.0
numpy==2.1.2
odfpy==1.4.1
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
tablib==3.7.0
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xlrd==2.0.1
xlwt==1.3.0
